+++ HospitalManagement/src/main/java/com/hospital/repository/SpecialityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface SpecialityRepository extends JpaRepository<Speciality, Integer> {
+public interface SpecialityRepository extends JpaRepository<Speciality, Long> {
     @Query("SELECT s FROM Speciality s WHERE s.position.positionId = :id")
-    public List<Speciality> findByPositionId(@Param("id") int position_id);
+    public List<Speciality> findByPositionId(@Param("id") Long position_id);
 }
